src/protocols/BLE/Adafruit_BLESniffer/requirements.txt
six>=1.11.0
pyserial>=3.4
